refactor: remove external tmux attach
Phase 0 of the tmux removal (#851): delete `tbd terminal attach`, the
app's "Copy Attach Command" item, the `terminal.attachCommand` RPC and
its shell recipe, the tmux socket-path resolver it alone consumed, and
the `tbd-ext-*` session reclamation arm together with the TmuxManager
session-listing and conditional-kill helpers that existed only for it.
The pty-holder transport design already declares external attach out of
scope.
