+++ Sources/TBDApp/TabBar.swift
@@ -2,9 +2,6 @@ import AppKit
 import SwiftUI
 import TBDShared
 import UniformTypeIdentifiers
-import os
-
-private let tabBarLogger = Logger(subsystem: "com.tbd.app", category: "tabBar")
 
 /// Transferable payload identifying a tab during drag/drop. App-only — never crosses the wire.
 struct TabDragPayload: Codable, Transferable {
@@ -580,16 +577,10 @@ enum ContinueInCodexMenu {
 
 // MARK: - TabTerminalTarget
 
-/// Pure resolution of the terminal a tab is backed by: the anchor for the
-/// tab's deep link, and the target of "Copy Attach Command". A pane with no
-/// terminal behind it (webview, code viewer, note) resolves to nil — its deep
-/// link falls back to the worktree alone, and it offers no attach command at
-/// all, because there is no tmux window for an external emulator to attach to.
+/// Pure resolution of the terminal a tab is backed by: the anchor its deep
+/// link points at. A pane with no terminal behind it (webview, code viewer,
+/// note) resolves to nil, and its deep link falls back to the worktree alone.
 ///
-/// Nil *is* the menu's hide condition: `contextMenuContent` gates the "Copy
-/// Attach Command" button on `if let` over this call, so there is deliberately
-/// no boolean twin — a separate `showsAttachCommand` would be a second rule
-/// the tests could hold green while the menu's own drifted away from it.
 /// Extracted from the view so each branch is unit-testable without rendering
 /// SwiftUI (same pattern as `TabParkMenuModel` and `ContinueInCodexMenu`).
 enum TabTerminalTarget {
@@ -954,35 +945,6 @@ private struct TabBarItem: View {
             )
         }
 
-        // Attach an external emulator (iTerm2, Terminal.app, Ghostty) to this
-        // tab's tmux window. The daemon composes the command: only it can
-        // resolve the socket path from the environment that created the tmux
-        // server, and only it verifies the pane's identity before naming the
-        // window. Nothing reaches the pasteboard unless that whole round trip
-        // succeeds — a half-formed command pasted into another terminal fails
-        // far from its cause.
-        if let attachTerminalID = TabTerminalTarget.terminalID(for: tab.content) {
-            Button("Copy Attach Command") {
-                let targetWorktreeID = worktreeID
-                Task {
-                    do {
-                        let result = try await appState.daemonClient.terminalAttachCommand(
-                            worktreeID: targetWorktreeID,
-                            terminalID: attachTerminalID
-                        )
-                        NSPasteboard.general.clearContents()
-                        NSPasteboard.general.setString(result.script, forType: .string)
-                    } catch {
-                        tabBarLogger.error(
-                            "terminal.attachCommand failed for terminal \(attachTerminalID, privacy: .public): \(error, privacy: .public)")
-                        appState.showAlert(
-                            "Couldn't copy the attach command: \(error.localizedDescription)",
-                            isError: true)
-                        appState.handleConnectionError(error)
-                    }
-                }
-            }
-        }
         Divider()
 
         if isClaudeTerminal {
