+++ Tests/TBDAppTests/TabTerminalTargetTests.swift
@@ -3,19 +3,15 @@ import Testing
 @testable import TBDApp
 import TBDShared
 
-/// Visibility of the tab context menu's "Copy Attach Command" item, and the
-/// terminal it targets. The item exists to hand the user a tmux attach command
-/// for *this tab's* terminal, so a tab with no terminal behind it must not
-/// offer it — there is no tmux window for an external emulator to attach to,
-/// and the daemon RPC it calls takes a terminal id that would not exist.
+/// Which terminal a tab's deep link is anchored to.
 ///
-/// The assertions go through `TabTerminalTarget.terminalID(for:)` because that
-/// is literally the menu's gate: `contextMenuContent` writes
-/// `if let attachTerminalID = TabTerminalTarget.terminalID(for: tab.content)`,
-/// so a nil here is the button's absence and a value is both its presence and
-/// the id it sends. There is no boolean twin to assert on — one would be a
-/// second rule these tests could hold green while the menu's own drifted.
-@Suite("Copy Attach Command menu visibility")
+/// `TabBarItem.linkTerminalID` is `TabTerminalTarget.terminalID(for:)`, and
+/// that value is what `DeepLink.makeShareableOpenURL` carries as the link's
+/// `terminal` component — so a nil here is a link that names the worktree
+/// alone, and a value is the tab the link reopens. Asserted through the
+/// resolver rather than a rendered menu, so each branch is testable without
+/// SwiftUI.
+@Suite("Tab deep-link terminal anchor")
 struct TabTerminalTargetTests {
     @Test("a terminal tab targets its own terminal")
     func terminalTabResolvesToItsTerminal() {
@@ -32,8 +28,8 @@ struct TabTerminalTargetTests {
         #expect(TabTerminalTarget.terminalID(for: content) == terminalID)
     }
 
-    @Test("tabs with no terminal behind them offer no attach command")
-    func nonTerminalTabsAreHidden() {
+    @Test("tabs with no terminal behind them anchor to no terminal")
+    func nonTerminalTabsAnchorToNothing() {
         let webview = PaneContent.webview(
             id: UUID(), url: URL(string: "https://example.com")!)
         let codeViewer = PaneContent.codeViewer(id: UUID(), path: "/tmp/file.swift")
